backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
anthropic==0.39.0
stripe==10.12.0
supabase==2.9.1
pydantic-settings==2.5.2
resend==2.4.0
twilio==9.3.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
